refactor(analysis-types): split derived DTOs
